+++ shardingsphere-error/shardingsphere-mysql-error/src/main/java/org/apache/shardingsphere/error/mysql/mapper/MySQLDialectExceptionMapper.java
@@ -21,15 +21,15 @@ import org.apache.shardingsphere.error.mapper.SQLDialectExceptionMapper;
 import org.apache.shardingsphere.error.mysql.code.MySQLVendorError;
 import org.apache.shardingsphere.error.vendor.ShardingSphereVendorError;
 import org.apache.shardingsphere.error.vendor.VendorError;
-import org.apache.shardingsphere.infra.exception.dialect.InsertColumnsAndValuesMismatchedException;
-import org.apache.shardingsphere.infra.exception.dialect.NoSuchTableException;
-import org.apache.shardingsphere.infra.exception.dialect.TableExistsException;
-import org.apache.shardingsphere.infra.exception.dialect.TableModifyInTransactionException;
-import org.apache.shardingsphere.infra.exception.dialect.TooManyConnectionsException;
-import org.apache.shardingsphere.infra.exception.dialect.database.DatabaseCreateExistsException;
-import org.apache.shardingsphere.infra.exception.dialect.database.DatabaseDropNotExistsException;
-import org.apache.shardingsphere.infra.exception.dialect.database.NoDatabaseSelectedException;
-import org.apache.shardingsphere.infra.exception.dialect.database.UnknownDatabaseException;
+import org.apache.shardingsphere.infra.exception.dialect.data.InsertColumnsAndValuesMismatchedException;
+import org.apache.shardingsphere.infra.exception.dialect.syntax.table.NoSuchTableException;
+import org.apache.shardingsphere.infra.exception.dialect.syntax.table.TableExistsException;
+import org.apache.shardingsphere.infra.exception.dialect.transaction.TableModifyInTransactionException;
+import org.apache.shardingsphere.infra.exception.dialect.connection.TooManyConnectionsException;
+import org.apache.shardingsphere.infra.exception.dialect.syntax.database.DatabaseCreateExistsException;
+import org.apache.shardingsphere.infra.exception.dialect.syntax.database.DatabaseDropNotExistsException;
+import org.apache.shardingsphere.infra.exception.dialect.syntax.database.NoDatabaseSelectedException;
+import org.apache.shardingsphere.infra.exception.dialect.syntax.database.UnknownDatabaseException;
 import org.apache.shardingsphere.infra.util.exception.inside.SQLDialectException;
 
 import java.sql.SQLException;
