+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/exception/dialect/syntax/database/DatabaseDropNotExistsException.java
@@ -15,20 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.exception.dialect;
+package org.apache.shardingsphere.infra.exception.dialect.syntax.database;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.infra.util.exception.inside.SQLDialectException;
 
 /**
- * Table exists exception.
+ * Database drop exists exception.
  */
 @RequiredArgsConstructor
 @Getter
-public final class TableExistsException extends SQLDialectException {
+public final class DatabaseDropNotExistsException extends SQLDialectException {
     
-    private static final long serialVersionUID = 6056681626545854214L;
+    private static final long serialVersionUID = 6088272565526510361L;
     
-    private final String tableName;
+    private final String databaseName;
 }
