+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/exception/dialect/syntax/database/NoDatabaseSelectedException.java
@@ -15,20 +15,14 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.exception.dialect;
+package org.apache.shardingsphere.infra.exception.dialect.syntax.database;
 
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.infra.util.exception.inside.SQLDialectException;
 
 /**
- * In transaction exception.
+ * No database selected exception.
  */
-@RequiredArgsConstructor
-@Getter
-public final class InTransactionException extends SQLDialectException {
+public final class NoDatabaseSelectedException extends SQLDialectException {
     
-    private static final long serialVersionUID = 8125852172126111520L;
-    
-    private final String message;
+    private static final long serialVersionUID = 7820981156141678308L;
 }
