+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/exception/dialect/syntax/database/UnknownDatabaseException.java
@@ -15,14 +15,20 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.infra.exception.dialect;
+package org.apache.shardingsphere.infra.exception.dialect.syntax.database;
 
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
 import org.apache.shardingsphere.infra.util.exception.inside.SQLDialectException;
 
 /**
- * Too many connections exception.
+ * Unknown database exception.
  */
-public final class TooManyConnectionsException extends SQLDialectException {
+@RequiredArgsConstructor
+@Getter
+public final class UnknownDatabaseException extends SQLDialectException {
     
-    private static final long serialVersionUID = -4397915988239251541L;
+    private static final long serialVersionUID = 6442120909016308638L;
+    
+    private final String databaseName;
 }
